+++ ea-application/app-api/build.gradle
@@ -3,6 +3,7 @@ dependencies {
     implementation project(':ea-domain')
     implementation project(':ea-infra')
     implementation project(':ea-common')
+    implementation project(':ea-security')
 
     //API 문서화 Swagger-UI
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2'
@@ -11,14 +12,14 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-aop'
     implementation 'org.springframework.boot:spring-boot-configuration-processor'
 
-    //Security 의존 주입
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    testImplementation 'org.springframework.security:spring-security-test'
+//    //Security 의존 주입
+//    implementation 'org.springframework.boot:spring-boot-starter-security'
+//    testImplementation 'org.springframework.security:spring-security-test'
 
-    //JWT
-    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-    implementation 'io.jsonwebtoken:jjwt-impl:0.11.5'
-    implementation 'io.jsonwebtoken:jjwt-jackson:0.11.5' // JSON 처리용
+//    //JWT
+//    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
+//    implementation 'io.jsonwebtoken:jjwt-impl:0.11.5'
+//    implementation 'io.jsonwebtoken:jjwt-jackson:0.11.5' // JSON 처리용
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
 }
