+++ ea-security/src/main/java/com/nbe2/security/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.nbe2.api.global.config;
+package com.nbe2.security.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,7 +13,7 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 
 import lombok.RequiredArgsConstructor;
 
-import com.nbe2.api.global.jwt.JwtProvider;
+import com.nbe2.security.utils.JwtProvider;
 
 @EnableWebSecurity
 @Configuration
