+++ ea-security/src/main/java/com/nbe2/security/exception/JwtErrorCode.java
@@ -1,4 +1,4 @@
-package com.nbe2.api.global.exception;
+package com.nbe2.security.exception;
 
 import static com.nbe2.common.constants.EAConstants.*;
