+++ ea-security/src/main/resources/application.yaml
@@ -1,0 +1,16 @@
+---
+spring:
+  config:
+    activate:
+      on-profile: local
+jwt:
+  screat-key: ${JWT_KEY}
+
+---
+spring:
+  config:
+    activate:
+      on-profile: dev
+
+jwt:
+  screat-key: ${JWT_KEY}
