+++ ea-security/src/test/java/com/nbe2/security/EaSecurityApplicationTests.java
@@ -1,0 +1,11 @@
+package com.nbe2.security;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EaSecurityApplicationTests {
+
+    @Test
+    void contextLoads() {}
+}
