+++ settings.gradle
@@ -5,3 +5,4 @@ include('ea-application:app-api')
 include('ea-domain')
 include('ea-infra')
 include('ea-common')
+include('ea-security')
